split_patch_rs_test.sh: proper quoting

+++ split_patch_rs_test.sh
@@ -11,25 +11,25 @@ DIR="/tmp"
 SHA=...
 
 PATCH_DIR=$DIR/PATCHES
-mkdir -p $PATCH_DIR
-rm -rf $PATCH_DIR/*
+mkdir -p "$PATCH_DIR"
+rm -rf "${PATCH_DIR:?}"/*
 
-git format-patch -o $PATCH_DIR $SHA 
+git format-patch -o "$PATCH_DIR" "$SHA"
 
 for opt in "hunks" "changes" "quiet" ; do
     echo "Working on option: $opt"
     pl_dir="$DIR/pl/$opt"
-    mkdir -p $pl_dir
-    rm -rf $pl_dir/*
+    mkdir -p "$pl_dir"
+    rm -rf "${pl_dir:?}"/*
     rs_dir="$DIR/rs/$opt"
-    mkdir -p $rs_dir
-    rm -rf $rs_dir/*
+    mkdir -p "$rs_dir"
+    rm -rf "${rs_dir:?}"/*
 
-    cp $PATCH_DIR/* $pl_dir
+    cp "$PATCH_DIR"/* "$pl_dir"
     "$SPLIT_PATCH_PL" "--$opt" "$pl_dir"/*
 
-    cp $PATCH_DIR/* $rs_dir
+    cp "$PATCH_DIR"/* "$rs_dir"
     "$SPLIT_PATCH_RS" "--$opt" "$rs_dir"/*
 
-    diff -ru $pl_dir $rs_dir
+    diff -ru "$pl_dir" "$rs_dir"
 done
